+++ SonarrSharp/Models/QualityQuality.cs
@@ -14,7 +14,7 @@ namespace SonarrSharp.Models
         /// <value>
         /// The identifier.
         /// </value>
-        [J("id")] public long Id { get; set; }
+        [J("id")] public int Id { get; set; }
 
         /// <summary>
         /// Gets or sets the name.
